Add profile creation and updating

+++ storage/models.go
@@ -18,23 +18,23 @@ func (s MediaStatus) String() string {
 type (
 	MediaStatus int
 	Media       struct {
-		ID           int64       `json:"id" db:"MediaID"`
+		ID           int         `json:"id" db:"MediaID"`
 		Name         string      `json:"name" db:"MediaName"`
 		Path         string      `json:"path" db:"Path"`
-		Duration     int64       `json:"duration" db:"Duration"`
-		LastModified int64       `json:"last_modified" db:"LastModified"`
+		Duration     int         `json:"duration" db:"Duration"`
+		LastModified int         `json:"last_modified" db:"LastModified"`
 		Status       MediaStatus `json:"status" db:"Status"`
 	}
 
 	Profile struct {
-		ID   int64  `json:"id" db:"ProfileID"`
+		ID   int    `json:"id" db:"ProfileID"`
 		Name string `json:"name" db:"ProfileName"`
 	}
 
 	ProfileViewingInfo struct {
-		MediaID   int64 `json:"media_id" db:"MediaID"`
-		ProfileID int64 `json:"profile_id" db:"ProfileID"`
-		TimePoint int64 `json:"time_point" db:"TimePoint"`
-		Timestamp int64 `json:"timestamp" db:"Timestamp"`
+		MediaID   int `json:"media_id" db:"MediaID"`
+		ProfileID int `json:"profile_id" db:"ProfileID"`
+		TimePoint int `json:"time_point" db:"TimePoint"`
+		Timestamp int `json:"timestamp" db:"Timestamp"`
 	}
 )
